update-image-profile muved to the correct route.
